requirements.txt
fastapi[standard]==0.110.2
yfinance==0.2.61
uvicorn==0.29.0
pydantic==2.7.1
scikit-learn==1.6.1
torch==2.7.0
pytorch-lightning==2.5.1.post0
boto3==1.38.27
python-dotenv==1.1.0
